refactor(daemon): extract proxy route to routes/proxy.ts
Move the inline /api/proxy/image endpoint to its own route file,
matching the pattern used by all other daemon routes. This ensures
the proxy is always available on daemon startup via app.route().

Prevents the issue where manual daemon restarts could miss
inline routes due to stale tsx cache.
